ipo_watch_bot/requirements.txt
requests
bs4
alpaca-py
twilio